refactor: Remove unused scripts for updating theme status and cleaning up Shopify themes
